03/pyproject.toml
[project]
name = "deng"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dagster>=1.9.10",
    "dagster-webserver>=1.9.10",
    "pandas>=2.2.3",
    "polars>=1.20.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=19.0.0",
    "sqlalchemy>=2.0.37",
    "gcsfs>=2025.2.0",
    "google-cloud-bigquery>=3.29.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.2",
    "testcontainers[postgres]>=4.9.1",
]

[tool.ruff]
line-length = 100
